# contracts/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VideoAgents",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/jobs": {
      "get": {
        "summary": "List Jobs",
        "operationId": "list_jobs_api_jobs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Job"
                  },
                  "type": "array",
                  "title": "Response List Jobs Api Jobs Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Job",
        "operationId": "create_job_api_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Brief"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "summary": "Get Job",
        "operationId": "get_job_api_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/draft": {
      "patch": {
        "summary": "Patch Draft",
        "operationId": "patch_draft_api_jobs__job_id__draft_patch",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/runs": {
      "post": {
        "summary": "Run",
        "operationId": "run_api_jobs__job_id__runs_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/cancel": {
      "post": {
        "summary": "Cancel",
        "operationId": "cancel_api_jobs__job_id__cancel_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/resume": {
      "post": {
        "summary": "Resume",
        "operationId": "resume_api_jobs__job_id__resume_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/assets": {
      "post": {
        "summary": "Upload",
        "operationId": "upload_api_jobs__job_id__assets_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_api_jobs__job_id__assets_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/alignment": {
      "post": {
        "summary": "Alignment",
        "operationId": "alignment_api_jobs__job_id__alignment_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlignmentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get Alignment",
        "operationId": "get_alignment_api_jobs__job_id__alignment_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "asset_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/events": {
      "get": {
        "summary": "Events",
        "operationId": "events_api_jobs__job_id__events_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "After"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/artifacts/{artifact_id}": {
      "get": {
        "summary": "Artifact",
        "operationId": "artifact_api_artifacts__artifact_id__get",
        "parameters": [
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artifact Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/settings": {
      "get": {
        "summary": "Get Settings",
        "operationId": "get_settings_api_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Patch Settings",
        "operationId": "patch_settings_api_settings_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/{provider}": {
      "get": {
        "summary": "Models",
        "operationId": "models_api_models__provider__get",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "codex_cli",
                "claude_code_cli"
              ],
              "type": "string",
              "title": "Provider"
            }
          },
          {
            "name": "refresh",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Refresh"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelCatalog"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Brief": {
        "properties": {
          "topic": {
            "type": "string",
            "maxLength": 2000,
            "title": "Topic",
            "default": ""
          },
          "script_text": {
            "type": "string",
            "maxLength": 30000,
            "title": "Script Text",
            "default": ""
          },
          "audience": {
            "type": "string",
            "maxLength": 500,
            "title": "Audience",
            "default": "普通观众"
          },
          "platform": {
            "type": "string",
            "maxLength": 100,
            "title": "Platform",
            "default": "抖音竖屏"
          },
          "usage": {
            "type": "string",
            "enum": [
              "personal",
              "commercial",
              "unspecified"
            ],
            "title": "Usage",
            "default": "unspecified"
          },
          "target_seconds": {
            "type": "number",
            "maximum": 1800.0,
            "minimum": 1.0,
            "title": "Target Seconds",
            "default": 60
          },
          "width": {
            "type": "integer",
            "maximum": 3840.0,
            "minimum": 240.0,
            "title": "Width",
            "default": 1080
          },
          "height": {
            "type": "integer",
            "maximum": 3840.0,
            "minimum": 240.0,
            "title": "Height",
            "default": 1920
          },
          "fps": {
            "type": "integer",
            "maximum": 60.0,
            "minimum": 15.0,
            "title": "Fps",
            "default": 30
          },
          "source_urls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 50,
            "title": "Source Urls"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Brief"
      },
...
      "DraftRequest": {
        "properties": {
          "base_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Base Revision"
          },
          "brief": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Brief"
              },
              {
                "type": "null"
              }
            ]
          },
          "script": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Script"
              },
              {
                "type": "null"
              }
            ]
          },
          "timeline": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Timeline"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "base_revision"
        ],
        "title": "DraftRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Job": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "QUEUED",
              "RUNNING",
              "NEEDS_INPUT",
              "NEEDS_HUMAN",
              "READY_FOR_PUBLISH",
              "REJECTED",
              "FAILED",
              "CANCELLED"
            ],
            "title": "Status"
          },
          "stage": {
            "type": "string",
            "enum": [
              "idle",
              "materials",
              "script",
              "voice",
              "director",
              "render",
              "review",
              "complete"
            ],
            "title": "Stage"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "progress": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "brief": {
            "$ref": "#/components/schemas/Brief"
          },
          "script": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Script"
              },
              {
                "type": "null"
              }
            ]
          },
          "script_discussion": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScriptDiscussion"
              },
              {
                "type": "null"
              }
            ]
          },
          "timeline": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Timeline"
              },
              {
                "type": "null"
              }
            ]
          },
          "assets": {
            "items": {
              "$ref": "#/components/schemas/Asset"
            },
            "type": "array",
            "title": "Assets"
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/Artifact"
            },
            "type": "array",
            "title": "Artifacts"
          },
          "review": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Review"
              },
              {
                "type": "null"
              }
            ]
          },
          "pending_input": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Input"
          },
          "latest_event_id": {
            "type": "integer",
            "title": "Latest Event Id",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "job_id",
          "revision",
          "status",
          "stage",
          "message",
          "created_at",
          "updated_at",
          "brief"
        ],
        "title": "Job"
      },
...
      "RunRequest": {
        "properties": {
          "base_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Base Revision"
          },
          "action": {
            "type": "string",
            "enum": [
              "produce",
              "voice",
              "storyboard",
              "preview",
              "final",
              "review"
            ],
            "title": "Action"
          },
          "idempotency_key": {
            "type": "string",
            "maxLength": 200,
            "minLength": 8,
            "title": "Idempotency Key"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "base_revision",
          "action",
          "idempotency_key"
        ],
        "title": "RunRequest"
      },
...
    }
  }
}